+++ cups/langprintf.c
@@ -4,7 +4,7 @@
  *   Localized printf/puts functions for the Common UNIX Printing
  *   System (CUPS).
  *
- *   Copyright 2002-2006 by Easy Software Products.
+ *   Copyright 2002-2007 by Easy Software Products.
  *
  *   These coded instructions, statements, and computer programs are the
  *   property of Easy Software Products and are protected by Federal
@@ -88,7 +88,7 @@ _cupsLangPrintf(FILE        *fp,	/* I - File to write to */
   */
 
   if (bytes > 0)
-    return (fwrite(output, 1, bytes, fp));
+    return ((int)fwrite(output, 1, bytes, fp));
   else
     return (bytes);
 }
@@ -133,7 +133,7 @@ _cupsLangPuts(FILE        *fp,		/* I - File to write to */
   */
 
   if (bytes > 0)
-    return (fwrite(output, 1, bytes, fp));
+    return ((int)fwrite(output, 1, bytes, fp));
   else
     return (bytes);
 }
