+++ cups/language.c
@@ -3,7 +3,7 @@
  *
  *   I18N/language support for the Common UNIX Printing System (CUPS).
  *
- *   Copyright 1997-2006 by Easy Software Products.
+ *   Copyright 1997-2007 by Easy Software Products.
  *
  *   These coded instructions, statements, and computer programs are the
  *   property of Easy Software Products and are protected by Federal
@@ -887,7 +887,7 @@ _cupsMessageLoad(const char *filename)	/* I - Message catalog to load */
       * Append to current string...
       */
 
-      length = strlen(m->str ? m->str : m->id);
+      length = (int)strlen(m->str ? m->str : m->id);
 
       if ((temp = realloc(m->str ? m->str : m->id,
                           length + strlen(ptr) + 1)) == NULL)
