+++ cups/string.c
@@ -3,7 +3,7 @@
  *
  *   String functions for the Common UNIX Printing System (CUPS).
  *
- *   Copyright 1997-2006 by Easy Software Products.
+ *   Copyright 1997-2007 by Easy Software Products.
  *
  *   These coded instructions, statements, and computer programs are the
  *   property of Easy Software Products and are protected by Federal
@@ -243,7 +243,7 @@ _cupsStrFormatd(char         *buf,	/* I - String */
   if (loc && loc->decimal_point)
   {
     dec    = loc->decimal_point;
-    declen = strlen(dec);
+    declen = (int)strlen(dec);
   }
   else
   {
